+++ Info/tutorial.csv
@@ -1,4 +1,4 @@
 씬번호,대사,완료시취해야하는행동,
-1,"안녕. 이건 튜토리얼이야",NOT,NXT,
-2,"아무 것도 하지 않고 종료할 거야",BWS,NXT,
+1,"안녕. 이건 튜토리얼이야",NOT,MVC,
+2,"아무 것도 하지 않고 종료할 거야",BWS,MVH,
 3,"튜토리얼을 완료했습니다",NOT,NXT,
