+++ Scripts/Tutorial/TutorialStep.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bbeb2e6f438bb164aa0dec88bb11e1e9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
